Renames "Resources" folder to "Resource"
This increases consistency with other folders, which are singular rather
than plural in regards to the folder contents.
