+++ NAS2D/NAS2D.vcxproj
@@ -198,12 +198,12 @@
     <ClCompile Include="Renderer\RectangleSkin.cpp" />
     <ClCompile Include="Renderer\Renderer.cpp" />
     <ClCompile Include="Renderer\RendererOpenGL.cpp" />
-    <ClCompile Include="Resources\AnimationSet.cpp" />
-    <ClCompile Include="Resources\Font.cpp" />
-    <ClCompile Include="Resources\Image.cpp" />
-    <ClCompile Include="Resources\Music.cpp" />
-    <ClCompile Include="Resources\Sound.cpp" />
-    <ClCompile Include="Resources\Sprite.cpp" />
+    <ClCompile Include="Resource\AnimationSet.cpp" />
+    <ClCompile Include="Resource\Font.cpp" />
+    <ClCompile Include="Resource\Image.cpp" />
+    <ClCompile Include="Resource\Music.cpp" />
+    <ClCompile Include="Resource\Sound.cpp" />
+    <ClCompile Include="Resource\Sprite.cpp" />
     <ClCompile Include="StateManager.cpp" />
     <ClCompile Include="StringUtils.cpp" />
     <ClCompile Include="Timer.cpp" />
@@ -252,13 +252,13 @@
     <ClInclude Include="Renderer\RectangleSkin.h" />
     <ClInclude Include="Renderer\Renderer.h" />
     <ClInclude Include="Renderer\RendererOpenGL.h" />
-    <ClInclude Include="Resources\ResourceCache.h" />
-    <ClInclude Include="Resources\AnimationSet.h" />
-    <ClInclude Include="Resources\Font.h" />
-    <ClInclude Include="Resources\Image.h" />
-    <ClInclude Include="Resources\Music.h" />
-    <ClInclude Include="Resources\Sound.h" />
-    <ClInclude Include="Resources\Sprite.h" />
+    <ClInclude Include="Resource\ResourceCache.h" />
+    <ClInclude Include="Resource\AnimationSet.h" />
+    <ClInclude Include="Resource\Font.h" />
+    <ClInclude Include="Resource\Image.h" />
+    <ClInclude Include="Resource\Music.h" />
+    <ClInclude Include="Resource\Sound.h" />
+    <ClInclude Include="Resource\Sprite.h" />
     <ClInclude Include="Signal/SignalConnection.h" />
     <ClInclude Include="Signal/Delegate.h" />
     <ClInclude Include="Signal/Signal.h" />
