+++ NAS2D/Resource/AnimationSet.h
