+++ NAS2D/Resource/Music.cpp
