+++ NAS2D/Resource/Sound.cpp
