+++ NAS2D/Resource/Sprite.cpp
