+++ NAS2D/Resource/Sprite.h
