+++ test/Resource/Image.test.cpp
@@ -1,0 +1,2 @@
+#include "NAS2D/Resource/Image.h"
+#include <gtest/gtest.h>
