--- test/Resources/Image.test.cpp
@@ -1,2 +1,0 @@
-#include "NAS2D/Resources/Image.h"
-#include <gtest/gtest.h>
