+++ test/test.vcxproj
@@ -51,8 +51,8 @@
     <ClCompile Include="Renderer/Rectangle.test.cpp" />
     <ClCompile Include="Renderer/Vector.test.cpp" />
     <ClCompile Include="Renderer/VectorSizeRange.test.cpp" />
-    <ClCompile Include="Resources/Image.test.cpp" />
-    <ClCompile Include="Resources/ResourceCache.test.cpp" />
+    <ClCompile Include="Resource/Image.test.cpp" />
+    <ClCompile Include="Resource/ResourceCache.test.cpp" />
     <ClCompile Include="ContainerUtils.test.cpp" />
     <ClCompile Include="Filesystem.test.cpp" />
     <ClCompile Include="MathUtils.test.cpp" />
